Move routes to top level
Minor cosmetic nit to avoid too nested structures
