+++ src/services/httpServer.ts
@@ -1,7 +1,7 @@
 import bodyParser from "body-parser";
 import express from "express";
 
-import routes from "./routes";
+import routes from "../routes";
 
 const app = express();
 const port = 3000;
